исправлены ошибки при мерже

+++ src/main/java/ru/valerii/NauJava/repository/AccountRepository.java
@@ -3,8 +3,9 @@ package ru.valerii.NauJava.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.valerii.NauJava.entity.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountRepository extends JpaRepository<Account, Long>, AccountRepositoryCustom {
-    List<Account> findByBalanceBetweenAndCurrency(Double minBalance, Double maxBalance, String currency);
+    List<Account> findByBalanceBetweenAndCurrency(BigDecimal minBalance, BigDecimal maxBalance, String currency);
 }
