+++ src/main/java/ru/valerii/NauJava/repository/AccountRepositoryCustom.java
@@ -1,8 +1,10 @@
 package ru.valerii.NauJava.repository;
 
 import ru.valerii.NauJava.entity.Account;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountRepositoryCustom {
-    List<Account> searchAccountsByCriteria(Double min, Double max, String currency);
+    List<Account> searchAccountsByCriteria(BigDecimal min, BigDecimal max, String currency);
 }
